Feat:ì카테고리 조회 수정 Setter제ê±

+++ src/main/java/com/ssgsakk/ssgdotcom/category/dto/CategoryDTO.java
@@ -1,5 +1,4 @@
 package com.ssgsakk.ssgdotcom.category.dto;
-import com.ssgsakk.ssgdotcom.category.domain.Category;
 import lombok.*;
 @Data
 @Builder
@@ -9,7 +8,7 @@ public class CategoryDTO {
 
     private String categoryName;
     private Long parentCategorySeq;
-    private Long level;
+    private int level;
 
 }
